kafka_3: refactor packages
