+++ src/main/java/sbp/school/kafka/common/utils/ChecksumHelper.java
@@ -1,4 +1,4 @@
-package sbp.school.kafka.utils;
+package sbp.school.kafka.common.utils;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
