+++ src/main/java/sbp/school/kafka/transaction/model/OperationType.java
@@ -1,4 +1,4 @@
-package sbp.school.kafka.transaction;
+package sbp.school.kafka.transaction.model;
 
 /**
  * Тип операции
