+++ src/main/java/sbp/school/kafka/transaction/utils/TransactionPartitioner.java
@@ -1,4 +1,4 @@
-package sbp.school.kafka.transaction;
+package sbp.school.kafka.transaction.utils;
 
 
 import org.apache.kafka.clients.producer.Partitioner;
@@ -6,6 +6,8 @@ import org.apache.kafka.common.Cluster;
 import org.apache.kafka.common.PartitionInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import sbp.school.kafka.transaction.model.OperationType;
+import sbp.school.kafka.transaction.model.TransactionDto;
 
 import java.util.List;
 import java.util.Map;
