+++ src/main/resources/kafka.properties
@@ -1,8 +1,8 @@
 # настройки производителя транзакций
 transaction.producer.bootstrap.servers=localhost:9092
 transaction.producer.key.serializer=org.apache.kafka.common.serialization.StringSerializer
-transaction.producer.value.serializer=sbp.school.kafka.transaction.TransactionSerializer
-transaction.producer.partitioner.class=sbp.school.kafka.transaction.TransactionPartitioner
+transaction.producer.value.serializer=sbp.school.kafka.transaction.utils.TransactionSerializer
+transaction.producer.partitioner.class=sbp.school.kafka.transaction.utils.TransactionPartitioner
 transaction.producer.acks=all
 transaction.producer.retries=2147483647
 transaction.producer.batch.size=100000
@@ -16,7 +16,7 @@ transaction.retry.period=PT5S
 # настройки потребителя подтверждений обработки транзакций
 transaction.ack.consumer.bootstrap.servers=localhost:9092
 transaction.ack.consumer.key.deserializer=org.apache.kafka.common.serialization.StringDeserializer
-transaction.ack.consumer.value.deserializer=sbp.school.kafka.ack.AckDeserializer
+transaction.ack.consumer.value.deserializer=sbp.school.kafka.ack.utils.AckDeserializer
 transaction.ack.consumer.group.id=transaction-ack-consumer
 transaction.ack.consumer.enable.auto.commit=false
 transaction.ack.consumer.auto.offset.reset=latest
